🔨 Refactor(#338): 내가 받은 리뷰 페이지 받은 카드, 한 줄 평 컴포넌트화
